# openapi/checklistix-data.yaml
openapi: 3.0.3
info:
  title: Checklistix Data API
  description: Checklistix Data API
  version: 1.0.0
servers:
  - url: 'https://localhost:9091'
    description: Development server
tags:
  - name: Checklists
paths:
  /checklists:
    get:
      tags:
        - Checklists
      summary: Returns a list of checklists
      responses:
        '200':
          description: An array of checklists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Checklist'
    post:
      tags:
        - Checklists
      summary: Creates a new checklist
      requestBody:
        description: Checklist information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistForm'
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /checklists/{id}:
    get:
      tags:
        - Checklists
      summary: Returns a single checklist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checklist'
        '404':
          description: Not found
        '500':
          description: Internal server error
    put:
      tags:
        - Checklists
      summary: Updates a checklist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Checklist information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistForm'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad request
        '500':
          description: Internal server error
    patch:
      tags:
        - Checklists
      summary: Partially updates a checklist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Checklist information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistForm'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad request
        '500':
          description: Internal server error
    delete:
      tags:
        - Checklists
      summary: Deletes a checklist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ok
        '204':
          description: No content
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /checklists/{id}/checklist-items:
    post:
      tags:
        - Checklists
      summary: Creates a new checklist item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Checklist item information
        required: true
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/SubChecklistForm'
                - $ref: '#/components/schemas/SectionTitleForm'
                - $ref: '#/components/schemas/TextBoxForm'
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /checklist-items/{id}:
    patch:
      tags:
        - Checklists
      summary: Updates a checklist item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Checklist item information
        required: true
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/SubChecklistForm'
                - $ref: '#/components/schemas/SectionTitleForm'
                - $ref: '#/components/schemas/TextBoxForm'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad request
        '500':
          description: Internal server error
    delete:
      tags:
        - Checklists
      summary: Deletes a checklist item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ok
        '204':
          description: No content
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /checklist-items/positions:
    patch:
      tags:
        - Checklists
      summary: Bulk update of checklist items
      requestBody:
        description: Checklist items
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PositionUpdate'
      responses:
        '204':
          description: Ok
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /checklist-items/{id}/sub-checklist-items:
    post:
      tags:
        - Checklists
      summary: Creates a new sub checklist item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: SubChecklist item information
        required: true
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/SubChecklistItemCheckItemForm'
                - $ref: '#/components/schemas/SubChecklistItemSubtitleForm'
                - $ref: '#/components/schemas/SubChecklistItemPreconditionForm'
                - $ref: '#/components/schemas/SubChecklistItemPostconditionForm'
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /sub-checklist-items/{id}:
    patch:
      tags:
        - Checklists
      summary: Updates a sub checklist item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Checklist item information
        required: true
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/SubChecklistItemCheckItemForm'
                - $ref: '#/components/schemas/SubChecklistItemSubtitleForm'
                - $ref: '#/components/schemas/SubChecklistItemPreconditionForm'
                - $ref: '#/components/schemas/SubChecklistItemPostconditionForm'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /sub-checklist-items/positions:
    patch:
      tags:
        - Checklists
      summary: Bulk update of sub checklist items
      requestBody:
        description: Sub checklist items
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PositionUpdate'
      responses:
        '204':
          description: Ok
        '400':
          description: Bad request
        '500':
          description: Internal server error

components:
  schemas:
    ChecklistForm:
      type: object
      properties:
        title:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        style:
          $ref: '#/components/schemas/ChecklistStyle'
        pageSize:
          $ref: '#/components/schemas/PageSize'
        pageOrientation:
          $ref: '#/components/schemas/PageOrientation'
        columns:
          type: integer
        fontSize:
          type: integer
        borderThickness:
          $ref: '#/components/schemas/BorderThickness'
        fontFamily:
          type: string
        defaultColor:
          type: string
        checklistItems:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/SubChecklist'
              - $ref: '#/components/schemas/SectionTitle'
              - $ref: '#/components/schemas/TextBox'

    Checklist:
      allOf:
        - $ref: '#/components/schemas/ChecklistForm'
        - type: object
          required:
            - id
            - created
            - style
            - pageSize
            - pageOrientation
            - columns
            - fontSize
            - borderThickness
            - fontFamily
            - defaultColor
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    PageSize:
      type: string
      enum:
        - A4
        - A5
        - Letter
        - HalfLetter

    PageOrientation:
      type: string
      enum:
        - Portrait
        - Landscape

    ChecklistStyle:
      type: string
      enum:
        - Dots
        - Line
        - MidDots
        - Dashed
    #        - Table

    BorderThickness:
      type: integer
      enum:
        - 0
        - 1
        - 2
        - 4
      x-enum-varnames:
        - None
        - Thin
        - Medium
        - Thick

    SubChecklistForm:
      type: object
      properties:
        checklistId:
          type: string
          format: uuid
          nullable: false
        column:
          type: integer
        page:
          type: integer
        position:
          type: integer
        title:
          type: string
        color:
          type: string
        subChecklistItems:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/SubChecklistItemCheckItem'
              - $ref: '#/components/schemas/SubChecklistItemSubtitle'
              - $ref: '#/components/schemas/SubChecklistItemPrecondition'
              - $ref: '#/components/schemas/SubChecklistItemPostcondition'
              - $ref: '#/components/schemas/SubChecklistItemLeftText'
              - $ref: '#/components/schemas/SubChecklistItemRightText'
        type:
          $ref: '#/components/schemas/ChecklistItemType' # always ChecklistItemType::SubChecklist
        # @TODO: Other props like to overwrite something like checklistStyle, borderThickness, ...

    SubChecklist:
      allOf:
        - $ref: '#/components/schemas/SubChecklistForm'
        - type: object
          required:
            - checklistId
            - column
            - page
            - position
            - id
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SectionTitleForm:
      type: object
      properties:
        checklistId:
          type: string
          format: uuid
          nullable: false
        column:
          type: integer
        page:
          type: integer
        position:
          type: integer
        text:
          type: string
        color:
          type: string
        type:
          $ref: '#/components/schemas/ChecklistItemType' # always ChecklistItemType::SectionTitle
        # @TODO: Other props like to overwrite something like checklistStyle, borderThickness, ...

    SectionTitle:
      allOf:
        - $ref: '#/components/schemas/SectionTitleForm'
        - type: object
          required:
            - id
            - checklistId
            - column
            - page
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    TextBoxForm:
      type: object
      properties:
        checklistId:
          type: string
          format: uuid
          nullable: false
        column:
          type: integer
        page:
          type: integer
        position:
          type: integer
        text:
          type: string
        color:
          type: string
        type:
          $ref: '#/components/schemas/ChecklistItemType' # always ChecklistItemType::TextBox
        # @TODO: Other props like to overwrite something like checklistStyle, borderThickness, ...

    TextBox:
      allOf:
        - $ref: '#/components/schemas/TextBoxForm'
        - type: object
          required:
            - id
            - checklistId
            - column
            - page
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    #    ColumnBreakForm:
    #      type: object
    #      properties:
    #        checklistId:
    #          type: string
    #          format: uuid
    #          nullable: false
    #        column:
    #          type: integer
    #        type:
    #          $ref: '#/components/schemas/ChecklistItemType' # always ChecklistItemType::ColumnBreak
    #
    #    ColumnBreak:
    #      allOf:
    #        - $ref: '#/components/schemas/ColumnBreakForm'
    #        - type: object
    #          properties:
    #            id:
    #              type: string
    #              format: uuid
    #              nullable: false

    #    PageBreakForm:
    #      type: object
    #      properties:
    #        checklistId:
    #          type: string
    #          format: uuid
    #          nullable: false
    #        type:
    #          $ref: '#/components/schemas/ChecklistItemType' # always ChecklistItemType::PageBreak
    #
    #    PageBreak:
    #      allOf:
    #        - $ref: '#/components/schemas/PageBreakForm'
    #        - type: object
    #          properties:
    #            id:
    #              type: string
    #              format: uuid
    #              nullable: false

    ChecklistItemType:
      type: string
      enum:
        - SubChecklist
        - SectionTitle
        - TextBox
    #        - ColumnBreak
    #        - PageBreak

    SubChecklistItemCheckItemForm:
      type: object
      properties:
        subChecklistId:
          type: string
          format: uuid
          nullable: false
        item:
          type: string
        action:
          type: string
        position:
          type: integer
        type:
          $ref: '#/components/schemas/SubChecklistItemType' # always SubChecklistItemType::CheckItem
        # @TODO: Other props like text decoration (bold, italic, underscore), ... (JSON?)

    SubChecklistItemCheckItem:
      allOf:
        - $ref: '#/components/schemas/SubChecklistItemCheckItemForm'
        - type: object
          required:
            - id
            - subChecklistId
            - type
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SubChecklistItemPreconditionForm:
      type: object
      properties:
        subChecklistId:
          type: string
          format: uuid
          nullable: false
        text:
          type: string
        position:
          type: integer
        type:
          $ref: '#/components/schemas/SubChecklistItemType' # always SubChecklistItemType::Precondition

    SubChecklistItemPrecondition:
      allOf:
        - $ref: '#/components/schemas/SubChecklistItemPreconditionForm'
        - type: object
          required:
            - id
            - subChecklistId
            - type
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SubChecklistItemPostconditionForm:
      type: object
      properties:
        subChecklistId:
          type: string
          format: uuid
          nullable: false
        text:
          type: string
        position:
          type: integer
        type:
          $ref: '#/components/schemas/SubChecklistItemType' # always SubChecklistItemType::Postcondition

    SubChecklistItemPostcondition:
      allOf:
        - $ref: '#/components/schemas/SubChecklistItemPostconditionForm'
        - type: object
          required:
            - id
            - subChecklistId
            - type
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SubChecklistItemSubtitleForm:
      type: object
      properties:
        subChecklistId:
          type: string
          format: uuid
          nullable: false
        text:
          type: string
        position:
          type: integer
        type:
          $ref: '#/components/schemas/SubChecklistItemType' # always SubChecklistItemType::Subtitle

    SubChecklistItemSubtitle:
      allOf:
        - $ref: '#/components/schemas/SubChecklistItemSubtitleForm'
        - type: object
          required:
            - id
            - subChecklistId
            - type
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SubChecklistItemLeftTextForm:
      type: object
      properties:
        subChecklistId:
          type: string
          format: uuid
          nullable: false
        text:
          type: string
        position:
          type: integer
        type:
          $ref: '#/components/schemas/SubChecklistItemType' # always SubChecklistItemType::Text

    SubChecklistItemLeftText:
      allOf:
        - $ref: '#/components/schemas/SubChecklistItemSubtitleForm'
        - type: object
          required:
            - id
            - subChecklistId
            - type
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SubChecklistItemRightTextForm:
      type: object
      properties:
        subChecklistId:
          type: string
          format: uuid
          nullable: false
        text:
          type: string
        position:
          type: integer
        type:
          $ref: '#/components/schemas/SubChecklistItemType' # always SubChecklistItemType::Text

    SubChecklistItemRightText:
      allOf:
        - $ref: '#/components/schemas/SubChecklistItemSubtitleForm'
        - type: object
          required:
            - id
            - subChecklistId
            - type
            - position
          properties:
            id:
              type: string
              format: uuid
              nullable: false

    SubChecklistItemType:
      type: string
      enum:
        - CheckItem
        - Precondition
        - Postcondition
        - Subtitle
        - LeftText
        - RightText

    PositionUpdate:
      type: object
      required:
        - id
        - position
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        position:
          type: integer